+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.1.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds check for returns
 - Adds `docsig.messages` to public scope
 - Ignores parameters named `_`